requirements.txt
Flask
mysql-connector-python
flask_socketio